mqtt-kafka/requirements.txt
confluent-kafka==2.0.2
paho-mqtt==1.6.1